# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SPMS API",
        "description": "subscription-platform-management-system API Documentation",
        "version": "API v1.0"
    },
    "paths": {
...
        "/api/v1/channels": {
            "get": {
                "tags": [
                    "Channels"
                ],
                "summary": "Get channels",
                "description": "Get Channels",
                "operationId": "ChannelsIndex",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Oauth2 token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer 1|iQ04fCi7gNIVCWnhgxubNJrdrIINlgnUkixEPfaA"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page.",
                        "required": false,
                        "schema": {
                            "type": "numeric"
                        },
                        "example": "5"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "schema": {
                            "type": "numeric"
                        },
                        "example": "1"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "HBO"
                    },
                    {
                        "name": "call_sign",
                        "in": "query",
                        "description": "call_sign",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "HBO"
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "active",
                        "required": false,
                        "schema": {
                            "type": "bool"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ChannelResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/channels/{id}": {
            "get": {
                "tags": [
                    "Channels"
                ],
                "summary": "Get Channel",
                "description": "Get channel by ID",
                "operationId": "GetChannel",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Oauth2 token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer 1|iQ04fCi7gNIVCWnhgxubNJrdrIINlgnUkixEPfaA"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Channel ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ChannelResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Channels"
                ],
                "summary": "Edit Channel",
                "operationId": "ChannelUpdate",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Oauth2 token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer 1|iQ04fCi7gNIVCWnhgxubNJrdrIINlgnUkixEPfaA"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Channel ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/ChannelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ChannelResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Channels"
                ],
                "summary": "Delete Channel",
                "operationId": "ChannelDelete",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Oauth2 token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer 1|iQ04fCi7gNIVCWnhgxubNJrdrIINlgnUkixEPfaA"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Channel ID",
                        "required": true,
                        "schema": {
                            "type": "numeric"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/channels/": {
            "post": {
                "tags": [
                    "Channels"
                ],
                "summary": "Create Channel",
                "operationId": "ChannelStore",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Oauth2 token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer 1|iQ04fCi7gNIVCWnhgxubNJrdrIINlgnUkixEPfaA"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/ChannelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ChannelResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ChannelRequest": {
                "required": [
                    "name",
                    "image_url",
                    "lineup_id",
                    "call_sign",
                    "active",
                    "display_order"
                ],
                "properties": {
                    "name": {
                        "description": "channel name",
                        "format": "string",
                        "example": "HBO",
                        "nullable": false
                    },
                    "image_url": {
                        "description": "image_url",
                        "format": "string",
                        "example": "https://via.placeholder.com/640x480.png/00aa22?text=illum",
                        "nullable": false
                    },
                    "lineup_id": {
                        "description": "lineup_id",
                        "format": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "call_sign": {
                        "description": "call_sign",
                        "format": "string",
                        "example": "HBO",
                        "nullable": false
                    },
                    "active": {
                        "description": "active",
                        "format": "boolean",
                        "example": "1",
                        "nullable": false
                    },
                    "display_order": {
                        "description": "display_order",
                        "format": "integer",
                        "example": "1",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
            "ChannelResource": {
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "format": "int64",
                        "default": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "TV channel name",
                        "format": "string",
                        "default": "HBO"
                    },
                    "image_url": {
                        "title": "image_url",
                        "description": "TV channel image url",
                        "format": "string",
                        "default": "https://via.placeholder.com/640x480.png/0011dd?text=eos"
                    },
                    "lineup_id": {
                        "title": "lineup_id",
                        "description": "TV channel lineup id",
                        "format": "int64",
                        "default": 1
                    },
                    "call_sign": {
                        "title": "call_sign",
                        "description": "TV channel call sign",
                        "format": "string",
                        "default": "HBO"
                    },
                    "active": {
                        "title": "active",
                        "description": "TV channel active",
                        "format": "boolean",
                        "default": true
                    },
                    "display_order": {
                        "title": "display_order",
                        "description": "TV channel display order",
                        "format": "int64",
                        "default": 1
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "TV channel created at",
                        "format": "string",
                        "default": "2024-01-01 00:00:00"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "TV channel updated at",
                        "format": "string",
                        "default": "2024-01-01 00:00:00"
                    },
                    "deleted_at": {
                        "title": "deleted_at",
                        "description": "TV channel deleted at",
                        "format": "string",
                        "default": "2024-01-01 00:00:00"
                    }
                },
                "type": "object"
            },
...
        }
    }
}